Fixed onClickListener()
JeansDetailsActivity is implementing View.OnClickListener (This could be
put in a separate class for separation of concerns - if you want one
class to deal with the onClickListener).

Refactored onClickHandler(View v) to onClick(View v) (The required
method for the implementation of View.OnClickListener).
Added log statements for testing (These can be deleted now).
Added break statements to break out of each case after it has run.

Updated activity_jeansdetails.xml so that all thumbnails use the onClick
onClickListener() defined in JeansDetailsActivity.